+++ packages/eslint-config/eslint.config.ts
@@ -1,8 +1,8 @@
-import { customRules } from "./rules/custom-rules";
-import { ignoredFolders } from "./rules/ignored-folders";
+import { customRules } from "./rules/custom";
+import { ignoredFolders } from "./rules/ignoredFolders";
 import { packageJsonRules } from "./rules/package-json-rules";
-import { storybookRules } from "./rules/storybook-rules";
-import { typescriptOverrideRules } from "./rules/typescript-override-rules";
+import { storybookRules } from "./rules/storybook";
+import { typescriptOverrideRules } from "./rules/typescriptOverrides";
 import eslintConfigCanonicalAuto from "eslint-config-canonical/auto";
 import tseslint from "typescript-eslint";
