+++ packages/eslint-config/rules/ignoredFolders.ts
